requirements.txt
fastapi
uvicorn
gspread_asyncio
google-auth
apscheduler
python-dotenv
python-multipart
httpx